AgregarProducto listo, si el identificador de la familia no existe entonces no se agrega el producto y se da el mensaje de error

+++ include/procedimientosBD.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void insertarProductoBD(char *id, char *nombre, char *familia, float costo, float precio, int cantidad_stock);
+void insertarProductoBD(char *id, char *nombre, char *familia, double costo, double precio, int cantidad_stock);
 void insertarFamiliaBD(char* idFamilia, char* descripcion);
 int conectar(MYSQL **conexion);
 #endif
